requirements.txt
black==24.8.0
entropy-estimators==0.0.1
faiss-gpu==1.7.2
hydra-core==1.3.2
matplotlib==3.8.4
mlflow==2.6.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==23.2
pacmap==0.7.0
pandas==2.2.0
pytorch==2.1.0
pytorch-lightning==2.2.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
tensorboard==2.17.0
torchvision==0.16.0
torchaudio==2.1.0
torchmetrics==1.8.2
tqdm==4.67.1
# ultralytics==8.2.1  # Install if using yolo evaluation or inference